gobject: Fix a few more -Wsign-conversion warnings
These are all fairly straightforward, but I didn’t get them locally;
they only showed up on CI.


Helps: #3405

+++ glib/gbsearcharray.h
@@ -182,7 +182,7 @@ g_bsearch_array_get_index (GBSearchArray        *barray,
                            const GBSearchConfig *bconfig,
                            gconstpointer         node_in_array)
 {
-  guint distance = ((guint8*) node_in_array) - G_BSEARCH_ARRAY_NODES (barray);
+  size_t distance = (size_t) (((guint8*) node_in_array) - G_BSEARCH_ARRAY_NODES (barray));
 
   g_return_val_if_fail (node_in_array != NULL, barray->n_nodes);
